managers: move regex into custom folder
