+++ lib/modules/manager/custom/api.ts
@@ -1,0 +1,7 @@
+import type { ManagerApi } from '../types';
+import * as regex from './regex';
+
+const api = new Map<string, ManagerApi>();
+export default api;
+
+api.set('regex', regex);
