+++ lib/modules/manager/custom/index.ts
@@ -1,0 +1,5 @@
+export {
+  defaultConfig,
+  extractPackageFile,
+  supportedDatasources,
+} from './regex';
