+++ lib/modules/manager/custom/regex/__fixtures__/ansible.yml
