+++ lib/modules/manager/custom/regex/__fixtures__/example.json
