+++ lib/modules/manager/custom/regex/__fixtures__/gitlab-ci.yml
