+++ lib/modules/manager/custom/regex/readme.md
