+++ lib/modules/manager/custom/regex/utils.ts
@@ -1,10 +1,10 @@
 import { URL } from 'node:url';
 import is from '@sindresorhus/is';
-import { migrateDatasource } from '../../../config/migrations/custom/datasource-migration';
-import type { RegexManagerTemplates } from '../../../config/types';
-import { logger } from '../../../logger';
-import * as template from '../../../util/template';
-import type { CustomExtractConfig, PackageDependency } from '../types';
+import { migrateDatasource } from '../../../../config/migrations/custom/datasource-migration';
+import type { RegexManagerTemplates } from '../../../../config/types';
+import { logger } from '../../../../logger';
+import * as template from '../../../../util/template';
+import type { CustomExtractConfig, PackageDependency } from '../../types';
 import type { ExtractionTemplate } from './types';
 
 export const validMatchFields = [
